实验复现/requirements.txt
numpy>=1.24
pandas>=2.0
scipy>=1.10
scikit-learn>=1.3
matplotlib>=3.7
PyYAML>=6.0
joblib>=1.3

